requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.16
django-cors-headers>=4.0
channels>=4.0
daphne>=4.0
dj-database-url>=2.0
psycopg2-binary>=2.9
whitenoise[brotli]>=6.0
gunicorn>=21.0
uvicorn[standard]>=0.30
